Object Detection Arduino Scripts
Two scripts: sketch_nov30b was the last one I tested and it works.
sketch_dec10a was functionalised to make for more simple combination with Joshu'as navigational system.

It returns two variables: activateLeft and activateRight

these each give the proportional activation for the brakes. Make sure to set the full activation and partial activation boundaries (ie. where brakes are full engaged and distances brakes start to engage respectively). The ratios are given from 0-1. 0 represents full activation (as in 0% of the current angular velocity) and 1 represents no activation (100% of the current angular velocity)